query-engine/requirements.txt
llama-index-core
llama-index-vector-stores-weaviate
llama-index-llms-openai
llama-index-llms-cohere
llama-index-embeddings-openai
llama-index-embeddings-cohere
cohere
openai
weaviate-client
pandas
fastapi
uvicorn
python-dotenv
pydantic
typing-inspect
typing-extensions
loguru
rich